Создает селектор для получения неповторяющихся городов
