+++ src/reducer/condition/selectors.js
@@ -10,16 +10,16 @@ export const getCity = (state) => {
   return state[NAME_SPACE].city;
 };
 
+export const getCities = (state) => {
+  return state[NAME_SPACE].cities;
+};
+
 export const getSortType = (state) => {
   return state[NAME_SPACE].sortType;
 };
 
 export const getOffers = (state) => {
   return state[NAME_SPACE].offers;
-};
-
-export const getCities = (state) => {
-  return state[NAME_SPACE].cities;
 };
 
 export const getFilteredOffers = createSelector(
@@ -37,5 +37,28 @@ export const getSortedOffers = createSelector(
     getFilteredOffers,
     (sortType, filteredOffers) => {
       return sortOffersByType(filteredOffers, sortType);
+    }
+);
+
+export const getNonRepeatingCities = createSelector(
+    getServerOffers,
+    (serverOffers) => {
+      const serverCities = serverOffers.map((offer) => {
+        return offer.city;
+      });
+
+      const nonRepeatingCities = [serverCities[0]];
+      const doesCityExist = (city) => {
+        return nonRepeatingCities.some((item) => {
+          return item.name === city.name;
+        });
+      };
+      serverCities.forEach((city) => {
+        if (!doesCityExist(city)) {
+          nonRepeatingCities.push(city);
+        }
+      });
+
+      return nonRepeatingCities;
     }
 );
